+++ CHANGELOG.md
@@ -121,6 +121,12 @@
   * Refactored `augmenters.pooling.MedianPooling` to use `imgaug.median_pool()`.
 * Added `imgaug.compute_paddings_to_reach_multiples_of()`.
 * Added `imgaug.pad_to_multiples_of()`.
+* Refactored `imgaug.pool()` to use `imgaug.pad()` for image padding.
+* [rarely breaking] Added a `pad_mode` argument to `imgaug.pool()`. This breaks
+  code relying on the order of the functions arguments.
+* Renamed argument `cval` to `pad_cval` in `imgaug.pool()`,
+  `imgaug.avg_pool()` and `imgaug.max_pool()`. The old name `cval` is now
+  deprecated.
 
 ## Fixes